requirements.txt
streamlit==1.28.1
streamlit-webrtc==0.47.0
opencv-python-headless==4.8.1.78
mysql-connector-python==8.2.0
pyngrok==7.0.4
av==10.0.0
pyaudio==0.2.13
schedule==1.2.0
pytz==2023.3
pygame==2.5.2
numpy==1.24.3
python-dotenv==1.0.0
bcrypt==4.1.1
redis==5.0.1
reportlab==4.0.7
requests==2.31.0
pillow==10.0.0
psutil==5.9.6
zoneinfo-backport==0.2.1
